+++ CPPLibrary/SimultaneousDiophantine.h
@@ -14,10 +14,10 @@ namespace CPPMathLibrary {
 	namespace SimultaneousDiophantine {
 
 		// Compute a Simultaneous Diophantine Approximation with a common divisor for all numerators. Assumes the vector x contains the desired accuracy of all real numbers.
-		__declspec(dllexport) QSMatrix<long> SameDivisor(const std::vector<double>& x, const double& alpha, const double& epsilon) throw (IncorrectDimensionException*);
+		__declspec(dllexport) QSMatrix<int> SameDivisor(const std::vector<double>& x, const double& alpha, const double& epsilon) throw (IncorrectDimensionException*);
 
 		// Compute a Simultaneous Diophantine Approximation with a common divisor for all numerators. Will compute a continued fraction approximation on the vector x.
-		__declspec(dllexport) QSMatrix<long> SameDivisorFromRealVector(const std::vector<double>& x, const double& alpha, const double& epsilon, const size_t& quotientCount) throw (IncorrectDimensionException*);
+		__declspec(dllexport) QSMatrix<int> SameDivisorFromRealVector(const std::vector<double>& x, const double& alpha, const double& epsilon, const size_t& quotientCount) throw (IncorrectDimensionException*);
 
 	}
